+++ Changelog.md
@@ -1,6 +1,11 @@
 # discord-chatgpt Changelog
 This file exists because sometimes I wait to commit my changes until I've done a bulk of work, but I still want to log exactly what's changing. These logs are ordered by date.
 
+## 2023-03-26
+- Add `cost` property to message dumps
+- Add the `/dalle` command
+    - Only `config.discord.owner_id` can use it for now
+
 ## 2023-03-22
 - Fix placeholder replacements
 - Store entire messages object in database